Fixing compiler warnings and follow style guide
More compiler warning fixes

Removed function hiding and interim defines

Fixed more compiler warnings

(cherry picked from commit 1be1587)

+++ datadriven/src/sgpp/datadriven/application/work_in_progress/RoundRobinScheduler.hpp
@@ -19,7 +19,7 @@ namespace sgpp {
             void assignTaskVariableTaskSize(TaskType taskType, AssignTaskResult &result) override;
 
         protected:
-            size_t lastWorkerID;
+            int lastWorkerID;
             size_t batchSize;
 
             size_t numOutstandingRequestsCurrentRefinement;
